feat: complete tree traversal and summarization, cleanup
